+++ websockets/client.py
@@ -98,7 +98,7 @@ class WebSocketClientProtocol(WebSocketCommonProtocol):
 
 @asyncio.coroutine
 def connect(uri, *,
-            loop=None, klass=WebSocketClientProtocol,
+            loop=None, klass=WebSocketClientProtocol, legacy_recv=False,
             origin=None, subprotocols=None, extra_headers=None,
             **kwds):
     """
@@ -136,7 +136,8 @@ def connect(uri, *,
         raise ValueError("connect() received a SSL context for a ws:// URI. "
                          "Use a wss:// URI to enable TLS.")
     factory = lambda: klass(
-        host=wsuri.host, port=wsuri.port, secure=wsuri.secure, loop=loop)
+        host=wsuri.host, port=wsuri.port, secure=wsuri.secure,
+        loop=loop, legacy_recv=legacy_recv)
 
     transport, protocol = yield from loop.create_connection(
         factory, wsuri.host, wsuri.port, **kwds)